streamlit/requirements.txt
streamlit==1.27.2
requests
kafka-python
pandas